分类表和消息表的映射关系及外键关联

+++ src/main/java/com/upublic/vo/Category.java
@@ -1,14 +1,14 @@
 package com.upublic.vo;
 
-public class Category {
-  private Long cid;
+public class  Category {
+  private int cid;
   private String cname;
 
-  public Long getCid() {
+  public int getCid() {
     return cid;
   }
 
-  public void setCid(Long cid) {
+  public void setCid(int cid) {
     this.cid = cid;
   }
